Cast before transmuting fn items to fix warnings

+++ src/internal/js/class.rs
@@ -249,7 +249,7 @@ pub trait ClassInternal: Class {
 
             let (call_callback, call_kernel) = match descriptor.call {
                 Some(k) => k.pair(),
-                None    => (mem::transmute(ConstructorCallKernel::unimplemented::<Self>), null_mut())
+                None    => (mem::transmute(ConstructorCallKernel::unimplemented::<Self> as usize), null_mut())
             };
 
             let metadata_pointer = neon_sys::class::create_base(isolate,
